🚀 feat: update staff data source to fetch from Sanity CMS
The staff data in the application was previously hardcoded in a JavaScript file. This commit updates the staff data source to fetch the data from Sanity CMS. The `staff` array has been replaced with a new function `fetchStaff` that makes a request to the Sanity API and returns the staff data. The fetched data is then assigned to the `staff` variable.